+++ src/components/listPage/cardBadge/CardBadge.tsx
@@ -1,4 +1,3 @@
-import styled from 'styled-components'
 import { DataType } from '../../../containers/listPage/CardList'
 import { Badge } from '../../../elements/Badge'
 import * as t from './CardBaged.style'
