+++ src/containers/listPage/CardList.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 //pages//
 import Card from '../../components/Card'
 import CardBadge from '../../components/listPage/cardBadge/CardBadge'
-import CardIcon from '../../components/listPage/CardIcon'
+import CardIcon from '../../components/listPage/cardIcon/CardIcon'
 import CateDrop from '../../components/listPage/CateDrop'
 import { theme } from '../../style/theme'
